+++ tests/settings.py
@@ -1,11 +1,12 @@
 HELPER_SETTINGS = {
     "SECRET_KEY": "djangocmsreferencestestsuitekey",
     "INSTALLED_APPS": [
+        "djangocms_alias.apps.AliasConfig",
         "djangocms_references",
         "djangocms_versioning",
         "djangocms_references.test_utils.app_1",
         "djangocms_references.test_utils.polls",
-        # "djangocms_references.test_utils.app_2",
+        "djangocms_references.test_utils.nested_references_app",
     ],
     "MIGRATION_MODULES": {
         "auth": None,
@@ -13,9 +14,11 @@ HELPER_SETTINGS = {
         "menus": None,
         "djangocms_references": None,
         "djangocms_versioning": None,
+        "djangocms_alias": None,
         "djangocms_references.test_utils.polls": None,
     },
     "DEFAULT_AUTO_FIELD": "django.db.models.AutoField",
+    "ROOT_URLCONF": "tests.urls",
 }