+++ tests/test_cms_config.py
@@ -10,6 +10,8 @@ from cms.models import PageContent
 from cms.toolbar.utils import get_object_preview_url
 from cms.utils.setup import configure_cms_apps
 
+from djangocms_alias.models import Alias
+
 from djangocms_references import cms_config
 from djangocms_references.test_utils import factories
 from djangocms_references.test_utils.app_1.models import Child, Parent
@@ -123,12 +125,16 @@ class ModifierTestCase(TestCase):
 
 class IntegrationTestCase(TestCase):
     def test_config_with_multiple_apps(self):
-        reference_models = apps.get_app_config(
+        cms_extension = apps.get_app_config(
             "djangocms_references"
-        ).cms_extension.reference_models
+        ).cms_extension
+        reference_models = cms_extension.reference_models
+        reference_plugins = cms_extension.reference_plugins
         expected_models = [Parent, Poll]
+        expected_plugins = [Alias, Poll]
 
         self.assertCountEqual(reference_models.keys(), expected_models)
+        self.assertCountEqual(reference_plugins.keys(), expected_plugins)
 
 
 class UnpublishDependenciesTestCase(TestCase):