+++ CHANGELOG.rst
@@ -4,6 +4,27 @@ Changelog
 
 Unreleased
 ==========
+* feat: Added filter by latest state
+
+1.2.0 (2022-04-14)
+==================
+* fix: Issue where a plugins name isn't the same as the registered plugin, treating a plugin as an object.
+* feat: Handle references objects that can be deeply nested through model relationships.
+
+1.1.1 (2022-04-11)
+==================
+* Fix: Added Alias list_action monkeypatch for references
+
+1.1.0 (2022-04-06)
+==================
+* Feature: Added Alias configuration
+
+1.0.1 (2022-03-31)
+==================
+* Fix: Added SVG manifest entry
+
+1.0.0 (2022-03-30)
+==================
 * Added list action icon template and svg
 * Python 3.8, 3.9 support added
 * Django 3.0, 3.1 and 3.2 support added